# api/src/swagger/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IToken": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"token": {
						"type": "string"
					},
					"createdToken": {
						"type": "number",
						"format": "double"
					},
					"expieredToken": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IPlatform": {
				"properties": {
					"idPlatformOfAdmin": {
						"type": "string"
					},
					"platformName": {
						"type": "string"
					},
					"platformId": {
						"type": "string"
					}
				},
				"required": [
					"idPlatformOfAdmin",
					"platformName",
					"platformId"
				],
				"type": "object"
			},
...
			"IUserType": {
				"enum": [
					"authAdmin",
					"Super User",
					"Simple User"
				],
				"type": "string"
			},
			"IPlatformList": {
				"properties": {
					"platformId": {
						"type": "string"
					},
					"platformName": {
						"type": "string"
					},
					"userProfile": {
						"properties": {
							"userProfileName": {
								"type": "string"
							},
							"userProfileBosConfigId": {
								"type": "string"
							},
							"userProfileAdminId": {
								"type": "string"
							},
							"userProfileId": {
								"type": "string"
							},
							"name": {
								"type": "string"
							}
						},
						"type": "object"
					}
				},
				"required": [
					"platformId",
					"userProfile"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUser": {
				"properties": {
					"id": {
						"anyOf": [
							{
								"type": "string"
							},
							{
								"type": "number",
								"format": "double"
							}
						]
					},
					"type": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"userName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"info": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/IUserType"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"platformList": {
						"items": {
							"$ref": "#/components/schemas/IPlatformList"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"userType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserCreationParams": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"info": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/IUserType"
					},
					"platformList": {
						"items": {
							"$ref": "#/components/schemas/IPlatformList"
						},
						"type": "array"
					}
				},
				"required": [
					"userName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserUpdateParams": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"info": {
						"type": "string"
					},
					"userType": {
						"$ref": "#/components/schemas/IUserType"
					},
					"grant_types": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"platformList": {
						"items": {
							"$ref": "#/components/schemas/IPlatformList"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"IUpdateUserPassword": {
				"properties": {
					"authAdminPassword": {
						"type": "string"
					},
					"userLastPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					}
				},
				"required": [
					"newPassword"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthAdminUpdateParams": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"telephone": {
						"type": "string"
					},
					"info": {
						"type": "string"
					}
				},
				"required": [
					"userName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserProfile": {
				"properties": {
					"userProfileName": {
						"type": "string"
					},
					"userProfileBosConfigId": {
						"type": "string"
					},
					"userProfileAdminId": {
						"type": "string"
					}
				},
				"required": [
					"userProfileName",
					"userProfileBosConfigId",
					"userProfileAdminId"
				],
				"type": "object"
			},
			"IUserPlatformProfile": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IPlatform"
					},
					{
						"properties": {
							"userProfile": {
								"$ref": "#/components/schemas/IUserProfile"
							}
						},
						"required": [
							"userProfile"
						],
						"type": "object"
					}
				]
			},
			"IUserToken": {
				"allOf": [
					{
						"$ref": "#/components/schemas/IToken"
					},
					{
						"properties": {
							"platformList": {
								"items": {
									"$ref": "#/components/schemas/IUserPlatformProfile"
								},
								"type": "array"
							},
							"serverId": {
								"type": "string"
							},
							"userProfile": {
								"type": "string"
							},
							"userType": {
								"type": "string"
							},
							"userId": {
								"type": "string"
							}
						},
						"type": "object"
					}
				]
			},
			"IUserLoginParams": {
				"properties": {
					"userName": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"userName",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IUserLogs": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"actor": {
						"properties": {
							"actorName": {
								"type": "string"
							},
							"actorId": {
								"type": "string"
							}
						},
						"required": [
							"actorName",
							"actorId"
						],
						"type": "object"
					}
				},
				"required": [
					"id",
					"name",
					"type",
					"date",
					"message",
					"actor"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "Spinal Organ Auth API",
		"version": "1.0.0",
		"description": "API for managing authentication and authorization in the Spinal Organ system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Spinalcom",
			"email": "developers@spinalcom.com"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserCreationParams"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IUser"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateAuthAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write",
							"ownData:write"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAuthAdminUpdateParams"
							}
						}
					}
				}
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read",
							"ownData:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"message": {
													"type": "string"
												},
												"error": {
													"type": "string"
												}
											},
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateParams"
							}
						}
					}
				}
			}
		},
		"/users/userInfo": {
			"post": {
				"operationId": "GetUserInfoByToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read",
							"ownData:read"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									}
								},
								"required": [
									"token"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/users/{userId}/updatePassword": {
			"put": {
				"operationId": "UpdateUserPassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUpdateUserPassword"
							}
						}
					}
				}
			}
		},
		"/users/getAuthAdmin": {
			"post": {
				"operationId": "GetAuthAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUser"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/userProfilesList": {
			"post": {
				"operationId": "UserProfilesList",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:write"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUserToken"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"all": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLoginParams"
							}
						}
					}
				}
			}
		},
		"/users/loginAuthAdmin": {
			"post": {
				"operationId": "LoginAuthAdmin",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUserToken"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"all": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLoginParams"
							}
						}
					}
				}
			}
		},
		"/users/getRoles": {
			"post": {
				"operationId": "GetRoles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"properties": {
													"name": {
														"type": "string"
													}
												},
												"required": [
													"name"
												],
												"type": "object"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"all": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}/userLogs": {
			"get": {
				"operationId": "GetUserLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"$ref": "#/components/schemas/IUserLogs"
											},
											"type": "array"
										},
										{
											"properties": {
												"error": {
													"type": "string"
												}
											},
											"required": [
												"error"
											],
											"type": "object"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": [
							"authAdmin:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}